+++ ReimuBase/RAPI.cs
@@ -108,7 +108,7 @@ namespace ReimuAPI.ReimuBase
             return false;
         }
 
-        public static bool getIsBotOP(int UserID)
+        public static bool getIsBotOP(long UserID)
         {
             ReimuConfig config = new ConfigManager().getConfig();
             foreach (int i in config.op_list)
